python-foxglove/pyproject.toml
[project]
name = "python-foxglove"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "foxglove-sdk>=0.16.1",
]
